added file dropping to ad mods
